requirements.txt
matplotlib==3.7.4
medmnist==2.2.3
numpy==1.24.4
pandas==2.0.3
Pillow==10.2.0
pytorch_ignite==0.4.13
scikit_learn==1.3.2
streamlit==1.30.0
torch==2.1.2
torchvision==0.16.2
monai==1.3.0